+++ examples/democoin/x/simplestake/msgs.go
@@ -12,12 +12,12 @@ import (
 
 // simple bond message
 type MsgBond struct {
-	Address sdk.Address   `json:"address"`
-	Stake   sdk.Coin      `json:"coins"`
-	PubKey  crypto.PubKey `json:"pub_key"`
+	Address sdk.AccAddress `json:"address"`
+	Stake   sdk.Coin       `json:"coins"`
+	PubKey  crypto.PubKey  `json:"pub_key"`
 }
 
-func NewMsgBond(addr sdk.Address, stake sdk.Coin, pubKey crypto.PubKey) MsgBond {
+func NewMsgBond(addr sdk.AccAddress, stake sdk.Coin, pubKey crypto.PubKey) MsgBond {
 	return MsgBond{
 		Address: addr,
 		Stake:   stake,
@@ -26,8 +26,8 @@ func NewMsgBond(addr sdk.Address, stake sdk.Coin, pubKey crypto.PubKey) MsgBond 
 }
 
 //nolint
-func (msg MsgBond) Type() string              { return moduleName } //TODO update "stake/createvalidator"
-func (msg MsgBond) GetSigners() []sdk.Address { return []sdk.Address{msg.Address} }
+func (msg MsgBond) Type() string                 { return moduleName } //TODO update "stake/createvalidator"
+func (msg MsgBond) GetSigners() []sdk.AccAddress { return []sdk.AccAddress{msg.Address} }
 
 // basic validation of the bond message
 func (msg MsgBond) ValidateBasic() sdk.Error {
@@ -55,19 +55,19 @@ func (msg MsgBond) GetSignBytes() []byte {
 
 // simple unbond message
 type MsgUnbond struct {
-	Address sdk.Address `json:"address"`
+	Address sdk.AccAddress `json:"address"`
 }
 
-func NewMsgUnbond(addr sdk.Address) MsgUnbond {
+func NewMsgUnbond(addr sdk.AccAddress) MsgUnbond {
 	return MsgUnbond{
 		Address: addr,
 	}
 }
 
 //nolint
-func (msg MsgUnbond) Type() string              { return moduleName } //TODO update "stake/createvalidator"
-func (msg MsgUnbond) GetSigners() []sdk.Address { return []sdk.Address{msg.Address} }
-func (msg MsgUnbond) ValidateBasic() sdk.Error  { return nil }
+func (msg MsgUnbond) Type() string                 { return moduleName } //TODO update "stake/createvalidator"
+func (msg MsgUnbond) GetSigners() []sdk.AccAddress { return []sdk.AccAddress{msg.Address} }
+func (msg MsgUnbond) ValidateBasic() sdk.Error     { return nil }
 
 // get unbond message sign bytes
 func (msg MsgUnbond) GetSignBytes() []byte {