+++ x/bank/app_test.go
@@ -18,12 +18,12 @@ import (
 // test bank module in a mock application
 var (
 	priv1     = crypto.GenPrivKeyEd25519()
-	addr1     = priv1.PubKey().Address()
+	addr1     = sdk.AccAddress(priv1.PubKey().Address())
 	priv2     = crypto.GenPrivKeyEd25519()
-	addr2     = priv2.PubKey().Address()
-	addr3     = crypto.GenPrivKeyEd25519().PubKey().Address()
+	addr2     = sdk.AccAddress(priv2.PubKey().Address())
+	addr3     = sdk.AccAddress(crypto.GenPrivKeyEd25519().PubKey().Address())
 	priv4     = crypto.GenPrivKeyEd25519()
-	addr4     = priv4.PubKey().Address()
+	addr4     = sdk.AccAddress(priv4.PubKey().Address())
 	coins     = sdk.Coins{sdk.NewCoin("foocoin", 10)}
 	halfCoins = sdk.Coins{sdk.NewCoin("foocoin", 5)}
 	manyCoins = sdk.Coins{sdk.NewCoin("foocoin", 1), sdk.NewCoin("barcoin", 1)}