+++ x/mock/test_utils.go
@@ -11,7 +11,7 @@ import (
 )
 
 // CheckBalance checks the balance of an account.
-func CheckBalance(t *testing.T, app *App, addr sdk.Address, exp sdk.Coins) {
+func CheckBalance(t *testing.T, app *App, addr sdk.AccAddress, exp sdk.Coins) {
 	ctxCheck := app.BaseApp.NewContext(true, abci.Header{})
 	res := app.AccountMapper.GetAccount(ctxCheck, addr)