+++ x/stake/client/rest/query.go
@@ -48,14 +48,14 @@ func delegationHandlerFn(ctx context.CoreContext, cdc *wire.Codec) http.HandlerF
 		bech32delegator := vars["delegator"]
 		bech32validator := vars["validator"]
 
-		delegatorAddr, err := sdk.GetAccAddressBech32(bech32delegator)
+		delegatorAddr, err := sdk.AccAddressFromBech32(bech32delegator)
 		if err != nil {
 			w.WriteHeader(http.StatusBadRequest)
 			w.Write([]byte(err.Error()))
 			return
 		}
 
-		validatorAddr, err := sdk.GetValAddressBech32(bech32validator)
+		validatorAddr, err := sdk.AccAddressFromBech32(bech32validator)
 		if err != nil {
 			w.WriteHeader(http.StatusBadRequest)
 			w.Write([]byte(err.Error()))
@@ -104,14 +104,14 @@ func ubdHandlerFn(ctx context.CoreContext, cdc *wire.Codec) http.HandlerFunc {
 		bech32delegator := vars["delegator"]
 		bech32validator := vars["validator"]
 
-		delegatorAddr, err := sdk.GetAccAddressBech32(bech32delegator)
+		delegatorAddr, err := sdk.AccAddressFromBech32(bech32delegator)
 		if err != nil {
 			w.WriteHeader(http.StatusBadRequest)
 			w.Write([]byte(err.Error()))
 			return
 		}
 
-		validatorAddr, err := sdk.GetValAddressBech32(bech32validator)
+		validatorAddr, err := sdk.AccAddressFromBech32(bech32validator)
 		if err != nil {
 			w.WriteHeader(http.StatusBadRequest)
 			w.Write([]byte(err.Error()))
@@ -161,21 +161,21 @@ func redHandlerFn(ctx context.CoreContext, cdc *wire.Codec) http.HandlerFunc {
 		bech32validatorSrc := vars["validator_src"]
 		bech32validatorDst := vars["validator_dst"]
 
-		delegatorAddr, err := sdk.GetAccAddressBech32(bech32delegator)
+		delegatorAddr, err := sdk.AccAddressFromBech32(bech32delegator)
 		if err != nil {
 			w.WriteHeader(http.StatusBadRequest)
 			w.Write([]byte(err.Error()))
 			return
 		}
 
-		validatorSrcAddr, err := sdk.GetValAddressBech32(bech32validatorSrc)
+		validatorSrcAddr, err := sdk.AccAddressFromBech32(bech32validatorSrc)
 		if err != nil {
 			w.WriteHeader(http.StatusBadRequest)
 			w.Write([]byte(err.Error()))
 			return
 		}
 
-		validatorDstAddr, err := sdk.GetValAddressBech32(bech32validatorDst)
+		validatorDstAddr, err := sdk.AccAddressFromBech32(bech32validatorDst)
 		if err != nil {
 			w.WriteHeader(http.StatusBadRequest)
 			w.Write([]byte(err.Error()))
@@ -217,9 +217,9 @@ func redHandlerFn(ctx context.CoreContext, cdc *wire.Codec) http.HandlerFunc {
 
 // TODO move exist next to validator struct for maintainability
 type StakeValidatorOutput struct {
-	Owner   string `json:"owner"`   // in bech32
-	PubKey  string `json:"pub_key"` // in bech32
-	Revoked bool   `json:"revoked"` // has the validator been revoked from bonded status?
+	Owner   sdk.AccAddress `json:"owner"`   // in bech32
+	PubKey  string         `json:"pub_key"` // in bech32
+	Revoked bool           `json:"revoked"` // has the validator been revoked from bonded status?
 
 	PoolShares      stake.PoolShares `json:"pool_shares"`      // total shares for tokens held in the pool
 	DelegatorShares sdk.Rat          `json:"delegator_shares"` // total shares issued to a validator's delegators
@@ -239,17 +239,13 @@ type StakeValidatorOutput struct {
 }
 
 func bech32StakeValidatorOutput(validator stake.Validator) (StakeValidatorOutput, error) {
-	bechOwner, err := sdk.Bech32ifyVal(validator.Owner)
-	if err != nil {
-		return StakeValidatorOutput{}, err
-	}
 	bechValPubkey, err := sdk.Bech32ifyValPub(validator.PubKey)
 	if err != nil {
 		return StakeValidatorOutput{}, err
 	}
 
 	return StakeValidatorOutput{
-		Owner:   bechOwner,
+		Owner:   validator.Owner,
 		PubKey:  bechValPubkey,
 		Revoked: validator.Revoked,