+++ x/stake/client/rest/tx.go
@@ -99,13 +99,13 @@ func editDelegationsRequestHandlerFn(cdc *wire.Codec, kb keys.Keybase, ctx conte
 
 		i := 0
 		for _, msg := range m.Delegations {
-			delegatorAddr, err := sdk.GetAccAddressBech32(msg.DelegatorAddr)
+			delegatorAddr, err := sdk.AccAddressFromBech32(msg.DelegatorAddr)
 			if err != nil {
 				w.WriteHeader(http.StatusInternalServerError)
 				w.Write([]byte(fmt.Sprintf("Couldn't decode delegator. Error: %s", err.Error())))
 				return
 			}
-			validatorAddr, err := sdk.GetValAddressBech32(msg.ValidatorAddr)
+			validatorAddr, err := sdk.AccAddressFromBech32(msg.ValidatorAddr)
 			if err != nil {
 				w.WriteHeader(http.StatusInternalServerError)
 				w.Write([]byte(fmt.Sprintf("Couldn't decode validator. Error: %s", err.Error())))
@@ -125,7 +125,7 @@ func editDelegationsRequestHandlerFn(cdc *wire.Codec, kb keys.Keybase, ctx conte
 		}
 
 		for _, msg := range m.BeginRedelegates {
-			delegatorAddr, err := sdk.GetAccAddressBech32(msg.DelegatorAddr)
+			delegatorAddr, err := sdk.AccAddressFromBech32(msg.DelegatorAddr)
 			if err != nil {
 				w.WriteHeader(http.StatusInternalServerError)
 				w.Write([]byte(fmt.Sprintf("Couldn't decode delegator. Error: %s", err.Error())))
@@ -136,13 +136,13 @@ func editDelegationsRequestHandlerFn(cdc *wire.Codec, kb keys.Keybase, ctx conte
 				w.Write([]byte("Must use own delegator address"))
 				return
 			}
-			validatorSrcAddr, err := sdk.GetValAddressBech32(msg.ValidatorSrcAddr)
+			validatorSrcAddr, err := sdk.AccAddressFromBech32(msg.ValidatorSrcAddr)
 			if err != nil {
 				w.WriteHeader(http.StatusInternalServerError)
 				w.Write([]byte(fmt.Sprintf("Couldn't decode validator. Error: %s", err.Error())))
 				return
 			}
-			validatorDstAddr, err := sdk.GetValAddressBech32(msg.ValidatorDstAddr)
+			validatorDstAddr, err := sdk.AccAddressFromBech32(msg.ValidatorDstAddr)
 			if err != nil {
 				w.WriteHeader(http.StatusInternalServerError)
 				w.Write([]byte(fmt.Sprintf("Couldn't decode validator. Error: %s", err.Error())))
@@ -164,19 +164,19 @@ func editDelegationsRequestHandlerFn(cdc *wire.Codec, kb keys.Keybase, ctx conte
 		}
 
 		for _, msg := range m.CompleteRedelegates {
-			delegatorAddr, err := sdk.GetAccAddressBech32(msg.DelegatorAddr)
+			delegatorAddr, err := sdk.AccAddressFromBech32(msg.DelegatorAddr)
 			if err != nil {
 				w.WriteHeader(http.StatusInternalServerError)
 				w.Write([]byte(fmt.Sprintf("Couldn't decode delegator. Error: %s", err.Error())))
 				return
 			}
-			validatorSrcAddr, err := sdk.GetValAddressBech32(msg.ValidatorSrcAddr)
+			validatorSrcAddr, err := sdk.AccAddressFromBech32(msg.ValidatorSrcAddr)
 			if err != nil {
 				w.WriteHeader(http.StatusInternalServerError)
 				w.Write([]byte(fmt.Sprintf("Couldn't decode validator. Error: %s", err.Error())))
 				return
 			}
-			validatorDstAddr, err := sdk.GetValAddressBech32(msg.ValidatorDstAddr)
+			validatorDstAddr, err := sdk.AccAddressFromBech32(msg.ValidatorDstAddr)
 			if err != nil {
 				w.WriteHeader(http.StatusInternalServerError)
 				w.Write([]byte(fmt.Sprintf("Couldn't decode validator. Error: %s", err.Error())))
@@ -196,7 +196,7 @@ func editDelegationsRequestHandlerFn(cdc *wire.Codec, kb keys.Keybase, ctx conte
 		}
 
 		for _, msg := range m.BeginUnbondings {
-			delegatorAddr, err := sdk.GetAccAddressBech32(msg.DelegatorAddr)
+			delegatorAddr, err := sdk.AccAddressFromBech32(msg.DelegatorAddr)
 			if err != nil {
 				w.WriteHeader(http.StatusInternalServerError)
 				w.Write([]byte(fmt.Sprintf("Couldn't decode delegator. Error: %s", err.Error())))
@@ -207,7 +207,7 @@ func editDelegationsRequestHandlerFn(cdc *wire.Codec, kb keys.Keybase, ctx conte
 				w.Write([]byte("Must use own delegator address"))
 				return
 			}
-			validatorAddr, err := sdk.GetValAddressBech32(msg.ValidatorAddr)
+			validatorAddr, err := sdk.AccAddressFromBech32(msg.ValidatorAddr)
 			if err != nil {
 				w.WriteHeader(http.StatusInternalServerError)
 				w.Write([]byte(fmt.Sprintf("Couldn't decode validator. Error: %s", err.Error())))
@@ -228,13 +228,13 @@ func editDelegationsRequestHandlerFn(cdc *wire.Codec, kb keys.Keybase, ctx conte
 		}
 
 		for _, msg := range m.CompleteUnbondings {
-			delegatorAddr, err := sdk.GetAccAddressBech32(msg.DelegatorAddr)
+			delegatorAddr, err := sdk.AccAddressFromBech32(msg.DelegatorAddr)
 			if err != nil {
 				w.WriteHeader(http.StatusInternalServerError)
 				w.Write([]byte(fmt.Sprintf("Couldn't decode delegator. Error: %s", err.Error())))
 				return
 			}
-			validatorAddr, err := sdk.GetValAddressBech32(msg.ValidatorAddr)
+			validatorAddr, err := sdk.AccAddressFromBech32(msg.ValidatorAddr)
 			if err != nil {
 				w.WriteHeader(http.StatusInternalServerError)
 				w.Write([]byte(fmt.Sprintf("Couldn't decode validator. Error: %s", err.Error())))